+++ todoman/__init__.py
@@ -1,4 +1,0 @@
-from todoman import version  # type: ignore
-
-__version__ = version.version
-__documentation__ = "https://todoman.rtfd.org/en/latest/"